packages/fl-manager-integrations/components/preprocessors/fl-manager-components-preprocessors-torchvision/pyproject.toml
[project]
name = "fl-manager-components-preprocessors-torchvision"
version = "0.1.0"
description = "fl-manager preprocessors torchvision integration"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.10"
authors = [
  { name = "Xin Lin", email = "xlin@itg.es" }
]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Science/Research",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Scientific/Engineering",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python",
  "Typing :: Typed",
  "Operating System :: OS Independent",
]
dependencies = [
  "fl-manager-core==0.*",
  "pillow>=11.2.1",
  "torch>=2.5.1",
  "torchvision>=0.19.1",
]

[project.urls]
Homepage = "https://github.com/public-itg/py-fl-manager"

[tool.uv.sources]
fl-manager-core = { workspace = true }

[tool.hatch.build.targets.wheel]
packages = ["src/fl_manager"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
